requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
secml-torch[foolbox,adv_lib]
numpy
foolbox